# src/IO.Swagger/Models/CalculationRequest.cs
/*
 * Secure Calculator API
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */
using System;
using System.Linq;
using System.IO;
using System.Text;
using System.Collections;
using System.Collections.Generic;
using System.Collections.ObjectModel;
using System.ComponentModel.DataAnnotations;
using System.Runtime.Serialization;
using Newtonsoft.Json;

namespace IO.Swagger.Models
{
    /// <summary>
    /// 
    /// </summary>
    [DataContract]
    public partial class CalculationRequest : IEquatable<CalculationRequest>
    { 
        /// <summary>
        /// Gets or Sets Operation
        /// </summary>
        [Required]

        [DataMember(Name="operation")]
        public OperationEnum Operation { get; set; }

        /// <summary>
        /// Gets or Sets Operand1
        /// </summary>
        [Required]

        [DataMember(Name="operand1")]
        public decimal? Operand1 { get; set; }

        /// <summary>
        /// Gets or Sets Operand2
        /// </summary>
        [Required]

        [DataMember(Name="operand2")]
        public decimal? Operand2 { get; set; }

        /// <summary>
        /// Returns the string presentation of the object
        /// </summary>
        /// <returns>String presentation of the object</returns>
        public override string ToString()
        {
            var sb = new StringBuilder();
            sb.Append("class CalculationRequest {\n");
            sb.Append("  Operation: ").Append(Operation).Append("\n");
            sb.Append("  Operand1: ").Append(Operand1).Append("\n");
            sb.Append("  Operand2: ").Append(Operand2).Append("\n");
            sb.Append("}\n");
            return sb.ToString();
        }

        /// <summary>
        /// Returns the JSON string presentation of the object
        /// </summary>
        /// <returns>JSON string presentation of the object</returns>
        public string ToJson()
        {
            return JsonConvert.SerializeObject(this, Formatting.Indented);
        }

        /// <summary>
        /// Returns true if objects are equal
        /// </summary>
        /// <param name="obj">Object to be compared</param>
        /// <returns>Boolean</returns>
        public override bool Equals(object obj)
        {
            if (ReferenceEquals(null, obj)) return false;
            if (ReferenceEquals(this, obj)) return true;
            return obj.GetType() == GetType() && Equals((CalculationRequest)obj);
        }

        /// <summary>
        /// Returns true if CalculationRequest instances are equal
        /// </summary>
        /// <param name="other">Instance of CalculationRequest to be compared</param>
        /// <returns>Boolean</returns>
        public bool Equals(CalculationRequest other)
        {
            if (ReferenceEquals(null, other)) return false;
            if (ReferenceEquals(this, other)) return true;

            return 
                (
                    Operation == other.Operation ||
                    Operation != null &&
                    Operation.Equals(other.Operation)
                ) && 
                (
                    Operand1 == other.Operand1 ||
                    Operand1 != null &&
                    Operand1.Equals(other.Operand1)
                ) && 
                (
                    Operand2 == other.Operand2 ||
                    Operand2 != null &&
                    Operand2.Equals(other.Operand2)
                );
        }

        /// <summary>
        /// Gets the hash code
        /// </summary>
        /// <returns>Hash code</returns>
        public override int GetHashCode()
        {
            unchecked // Overflow is fine, just wrap
            {
                var hashCode = 41;
                // Suitable nullity checks etc, of course :)
                    if (Operation != null)
                    hashCode = hashCode * 59 + Operation.GetHashCode();
                    if (Operand1 != null)
                    hashCode = hashCode * 59 + Operand1.GetHashCode();
                    if (Operand2 != null)
                    hashCode = hashCode * 59 + Operand2.GetHashCode();
                return hashCode;
            }
        }

        #region Operators
        #pragma warning disable 1591

        public static bool operator ==(CalculationRequest left, CalculationRequest right)
        {
            return Equals(left, right);
        }

        public static bool operator !=(CalculationRequest left, CalculationRequest right)
        {
            return !Equals(left, right);
        }

        #pragma warning restore 1591
        #endregion Operators
    }
}