# src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "Secure Calculator API",
    "termsOfService" : "https://example.com/terms",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "http://localhost:5000",
    "description" : "Local Development Server"
  } ],
  "paths" : {
    "/calculate" : {
      "post" : {
        "tags" : [ "Calculator" ],
        "summary" : "Perform a calculation",
        "requestBody" : {
          "description" : "Calculation parameters",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CalculationRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Calculation result",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CalculationResult"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal Server Error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "BearerAuth" : [ ]
        } ]
      }
    },
    "/auth/login" : {
      "post" : {
        "tags" : [ "Authentication" ],
        "summary" : "User login",
        "requestBody" : {
          "description" : "User credentials",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/User"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "JWT token",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuthToken"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal Server Error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/register" : {
      "post" : {
        "tags" : [ "Authentication" ],
        "summary" : "Register a new user",
        "requestBody" : {
          "description" : "New user data",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/User"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "User created successfully",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/User"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal Server Error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/logout" : {
      "delete" : {
        "tags" : [ "Authentication" ],
        "summary" : "Logout (delete user)",
        "parameters" : [ {
          "name" : "email",
          "in" : "query",
          "required" : true,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "User deleted successfully"
          },
          "400" : {
            "description" : "Bad Request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal Server Error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "AuthToken" : {
        "type" : "object",
        "properties" : {
          "token" : {
            "type" : "string"
          }
        },
        "example" : {
          "token" : "token"
        }
      },
      "CalculationRequest" : {
        "required" : [ "operand1", "operand2", "operation" ],
        "type" : "object",
        "properties" : {
          "operation" : {
            "$ref" : "#/components/schemas/OperationEnum"
          },
          "operand1" : {
            "type" : "number"
          },
          "operand2" : {
            "type" : "number"
          }
        }
      },
      "OperationEnum" : {
        "type" : "string",
        "enum" : [ "add", "subtract", "multiply", "divide" ]
      },
      "CalculationResult" : {
        "type" : "object",
        "properties" : {
          "result" : {
            "type" : "number"
          }
        },
        "example" : {
          "result" : 0.8008281904610115
        }
      },
      "ErrorResponse" : {
        "type" : "object",
        "properties" : {
          "error" : {
            "type" : "string",
            "description" : "Error message"
          }
        }
      },
      "User" : {
        "required" : [ "email" ],
        "type" : "object",
        "properties" : {
          "email" : {
            "type" : "string"
          }
        },
        "example" : {
          "email" : "email"
        }
      }
    },
    "securitySchemes" : {
      "BearerAuth" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}